Extract grid methods to a different file
